requirements.txt
# CORE FRAMEWORK - VERSIONES COMPATIBLES PROBADAS
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
# DATABASE
sqlalchemy==2.0.21
pyodbc==5.0.1
alembic==1.12.1
# CACHE - SOLO REDIS SINCRÓNICO
redis==5.0.1
# NOTA: NO incluir aioredis
# MACHINE LEARNING
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
# NLP AVANZADO (OPCIONAL)
transformers==4.33.3
torch==2.0.1
sentence-transformers==2.2.2
# OPENAI
openai==1.3.7
# UTILITIES
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
# EXCEL/CSV PROCESSING
openpyxl==3.1.2
xlsxwriter==3.1.9
tabulate==0.9.0
# SCHEDULER Y BACKGROUND TASKS
apscheduler==3.10.4
# MONITORING Y LOGGING
prometheus-client==0.17.1
structlog==23.1.0
# TESTING (DESARROLLO)
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
# DEVELOPMENT TOOLS
black==23.7.0
flake8==6.0.0
mypy==1.5.1
# YAML PROCESSING
PyYAML==6.0.1
# WINDOWS COMPATIBILITY
colorama==0.4.6
wmi==1.5.1; sys_platform == "win32"
# SECURITY
cryptography==41.0.4
bcrypt==4.0.1
# RATE LIMITING
slowapi==0.1.9
# STARLETTE (Compatible con FastAPI)
starlette==0.27.0
# TYPING EXTENSIONS
typing-extensions==4.7.1